API Master/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic
pydantic-settings
python-dotenv
websockets
